Added Db Support

+++ src/main/java/dev/app/bootiful/book/dto/Book.java
@@ -3,4 +3,4 @@ package dev.app.bootiful.book.dto;
 import java.time.LocalDate;
 
 public record Book(
-    Integer id, String title, String author, String description, String publishedDate) {}
+    Long id, String title, String author, String description, String publishedDate) {}
